+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added 
 
+- **[Possibly breaking change]** : Exception raised if both EntityTypeCode and EntityLogicalName are set when registering a plugin step. EntityTypeCode is deprecated, please move all plugin registrations to use EntityLogicalName instead.
+- Introduced new user-defined exceptions to verify plugin step registrations when EntityTypeCode is used
 - Adding support for secure and unsecure configurations in pipeline simulation - https://github.com/DynamicsValue/fake-xrm-easy/issues/103
 
 ### Changed